Cleaned up user, mentor, and internship data functions and reorganized folders. Added placeholder code to connect user/mentor authenicaiton with data storage (ensuring UIDs associated with sign-ups conect with UIDs of signed up user/mentor)
