+++ bigbyte/package.json
@@ -11,6 +11,7 @@
     "cors": "^2.8.5",
     "express": "^4.18.2",
     "firebase": "^10.8.0",
+    "firebase-admin": "^12.0.0",
     "nodemon": "^3.0.3",
     "react": "^18.2.0",
     "react-bootstrap": "^2.10.1",
