+++ bigbyte/src/pages/BackendFunctionality/dataManipulation/firebaseConfiguration.js
@@ -3,6 +3,7 @@ import { initializeApp } from "firebase/app";
 import { getAuth, GoogleAuthProvider } from "firebase/auth";
 import { getFirestore } from "firebase/firestore";
 import { getStorage } from "firebase/storage";
+
 export function getFirebaseConfig()
 {
     const firebaseConfig = {
